moved "get_cell" to utils
